Extract shared geodescriber sentence/params parsing code into the util file
